requirements.txt
# Core libraries
numpy>=1.20.0
pandas>=1.3.0
matplotlib>=3.4.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.62.0

# Deep learning
torch>=1.10.0
torchvision>=0.11.0

# Data processing
opencv-python>=4.5.0
Pillow>=8.0.0
mplfinance>=0.12.0

# Database
sqlite3>=2.6.0

# HTTP requests
requests>=2.26.0

# Environment management
python-dotenv>=0.19.0

# For time series data
statsmodels>=0.13.0 